refactor: LottoMachine 당첨통계 LottoCalculator 로 역할 위임
